refactor(frontend): PR-04 centralize API contract types for upcoming phases
Introduce shared service-level API types in api-types.ts and align admin/notification/knowledge services to consume and re-export centralized contracts without changing runtime behavior.

Made-with: Cursor
