+++ frontend/src/app/services/admin-audit.service.ts
@@ -2,22 +2,14 @@ import { Injectable } from '@angular/core';
 import { HttpClient } from '@angular/common/http';
 import { Observable } from 'rxjs';
 import { ApiContract } from './api-contract';
-
-export interface AuditLogEntry {
-  id: string;
-  actor_user_id: string;
-  action: string;
-  entity_type: string;
-  entity_id: string;
-  detail: string;
-  created_at: string;
-}
+import type { ApiListResponse, AuditLogEntry } from './api-types';
+export type { AuditLogEntry } from './api-types';
 
 @Injectable({ providedIn: 'root' })
 export class AdminAuditService {
   constructor(private http: HttpClient) {}
 
-  list(): Observable<{ entries: AuditLogEntry[] }> {
-    return this.http.get<{ entries: AuditLogEntry[] }>(ApiContract.admin.auditLog);
+  list(): Observable<ApiListResponse<AuditLogEntry, 'entries'>> {
+    return this.http.get<ApiListResponse<AuditLogEntry, 'entries'>>(ApiContract.admin.auditLog);
   }
 }
